refactor[api]: changes type of param

+++ stamp/src/main/java/com/accessibility/stamp/repository/StampRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface StampRepository extends JpaRepository<StampEntity, Long> {
-    StampEntity findByStampLevel(Float stampLevel);
+    StampEntity findByStampLevel(Integer stampLevel);
 }
